site: included logging-related flags, cleanup
